add filtering of logdate and tasklog in viewmodel

+++ app/src/main/java/com/kklabs/karam/data/DataSource.kt
@@ -27,7 +27,7 @@ interface DataSource {
     suspend fun createTasklog(request: CreateTasklogRequest): NetworkResponse<TasklogResponse>
 
     suspend fun getTasklogs(
-        taskId: Long,
+        taskId: Int,
         pageKey: Int? = null
     ): NetworkResponse<DataResponse<List<ModuleData<*>>>>
